Moving click handling to alert component from container in order to handle index correctly.  Removed AlertsComponent and added new AlertComponent.
